Change data to decimal (no functional changes), print out cycles (based on info from /proc/cpuinfo) - since so many people measure these benchmarks in cycles, rather than time.
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@7708 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ user/blkfin-apps/edn/edn.c
@@ -23,7 +23,7 @@ void vec_mpy1(short y[], const short x[], short scaler)
 /******************************************************
  *	Dot Product                                   *
  ******************************************************/
-int mac(const short *a, const short *b, long int sqr, long int *sum)
+int mac(const short *a, const short *b, int sqr, int *sum)
 {
 	int i;
 	int dotp = *sum;
